Finishing up the message routes and contact routes
