+++ src/FaxCoreRestClient/Models/Enumerators/AddressType.cs
@@ -3,7 +3,7 @@ using System.Text.Json.Serialization;
 namespace FaxCoreRestClient.Models.Enumerators
 {
     [JsonConverter(typeof(JsonStringEnumConverter))]
-    public enum PreferredAddressType
+    public enum AddressType
     {
         Email,
         Fax,
