+++ src/FaxCoreRestClient/Models/Enumerators/Visibility.cs
@@ -1,0 +1,10 @@
+namespace FaxCoreRestClient.Models.Enumerators
+{
+    public enum Visibility
+    {
+        Private,
+        Public,
+        Global,
+        Disabled
+    }
+}
