+++ src/FaxCoreRestClient/Models/Request/SendMessageRequest.cs
@@ -112,133 +112,4 @@ namespace FaxCoreRestClient.Models.Request
         [JsonPropertyName("agents")]
         public List<Agent> Agents { get; set; }
     }
-
-    public class InternalRecipient
-    {
-        /// <summary>
-        ///     The username to deliver the message to
-        /// </summary>
-        [DataMember(Name = "userName")]
-        [JsonPropertyName("userName")]
-        public string UserName { get; set; }
-    }
-
-    /// <summary>
-    ///     Represents a recipient object.
-    /// </summary>
-    [DataContract]
-    public class Recipient
-    {
-        /// <summary>
-        ///     Gets or sets the name of the recipient.
-        /// </summary>
-        [DataMember(Name = "name")]
-        [JsonPropertyName("name")]
-        public string Name { get; set; }
-
-        /// <summary>
-        ///     Gets or sets the address of the recipient.
-        /// </summary>
-        [DataMember(Name = "address")]
-        [JsonPropertyName("address")]
-        public string Address { get; set; }
-
-        /// <summary>
-        ///     Set whether the fax number should be dialed as is without any formatting. (Not valid for email)
-        /// </summary>
-        [DataMember(Name = "rawFax")]
-        [JsonPropertyName("rawFax")]
-        public bool IsRawFax { get; set; }
-
-        /// <summary>
-        ///     Gets or sets the notification address of the recipient.
-        /// </summary>
-        [DataMember(Name = "notifyAddress")]
-        [JsonPropertyName("notifyAddress")]
-        public string NotifyAddress { get; set; }
-
-        /// <summary>
-        ///     Gets or sets the company of the recipient.
-        /// </summary>
-        [DataMember(Name = "company")]
-        [JsonPropertyName("company")]
-        public string Company { get; set; }
-    }
-
-    /// <summary>
-    ///     Represents a tracking object.
-    /// </summary>
-    [DataContract]
-    public class Tracking
-    {
-        /// <summary>
-        ///     Gets or sets the label of the tracking.
-        /// </summary>
-        [DataMember(Name = "label")]
-        [JsonPropertyName("label")]
-        public string Label { get; set; }
-
-        /// <summary>
-        ///     Gets or sets the value of the tracking.
-        /// </summary>
-        [DataMember(Name = "value")]
-        [JsonPropertyName("value")]
-        public string Value { get; set; }
-    }
-
-    /// <summary>
-    ///     Represents a document object.
-    /// </summary>
-    [DataContract]
-    public class Document
-    {
-        /// <summary>
-        ///     Gets or sets the name of the document.
-        /// </summary>
-        [DataMember(Name = "name")]
-        [JsonPropertyName("name")]
-        public string Name { get; set; }
-
-        /// <summary>
-        ///     Gets or sets the path of the document.
-        /// </summary>
-        [DataMember(Name = "path")]
-        [JsonPropertyName("path")]
-        public string Path { get; set; }
-
-        /// <summary>
-        ///     Gets or sets the value indicating whether the document is a merge document.
-        /// </summary>
-        [DataMember(Name = "isMerge")]
-        [JsonPropertyName("isMerge")]
-        public bool IsMerge { get; set; }
-    }
-
-    /// <summary>
-    ///     Represents an agent object.
-    /// </summary>
-    [DataContract]
-    public class Agent
-    {
-        /// <summary>
-        ///     Gets or sets the ID of the agent.
-        /// </summary>
-        [DataMember(Name = "id")]
-        [JsonPropertyName("id")]
-        public string ID { get; set; }
-
-        /// <summary>
-        ///     Gets or sets the type of the agent.
-        /// </summary>
-        [DataMember(Name = "type")]
-        [JsonPropertyName("type")]
-        public string Type { get; set; }
-
-        /// <summary>
-        ///     Gets or sets the value of the agent.
-        /// </summary>
-        [DataMember(Name = "value")]
-        [JsonPropertyName("value")]
-        public string Value { get; set; }
-    }
 }
