+++ src/FaxCoreRestClient/Models/Request/UserConfigRequest.cs
@@ -14,22 +14,4 @@ namespace FaxCoreRestClient.Models.Request
         [JsonPropertyName("config")]
         public List<UserConfig> Config { get; set; }
     }
-
-    /// <summary>
-    ///     A User Config to set
-    /// </summary>
-    public class UserConfig
-    {
-        /// <summary>
-        ///     The name of the config to set
-        /// </summary>
-        [JsonPropertyName("configName")]
-        public string ConfigName { get; set; }
-
-        /// <summary>
-        ///     The value of the config to set
-        /// </summary>
-        [JsonPropertyName("configValue")]
-        public string ConfigValue { get; set; }
-    }
 }
