+++ src/FaxCoreRestClient/Models/Response/ContactCreationResponse.cs
@@ -1,0 +1,13 @@
+namespace FaxCoreRestClient.Models.Response
+{
+    /// <summary>
+    ///     The response from a contact creation request.
+    /// </summary>
+    public class ContactCreationResponse
+    {
+        /// <summary>
+        ///     The unique identifier for the contact.
+        /// </summary>
+        public int ContactId { get; set; }
+    }
+}
